backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
yfinance==0.2.31
websockets==11.0.3
python-dotenv==1.0.0